+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Navigation Changelog
 
+## 0.3.3
+
+* `MVVM`:
+  * New `ViewFragment#onSetNode` hook
+
 ## 0.3.2
 
 * `Versions`: